Refactor MCQEditionPanel.jsx
